Move view related files to sub folder view.
