+++ sample/src/main/res/layout/activity_main.xml
@@ -13,7 +13,7 @@
                 android:id="@+id/textView"
                 android:text="Hello"
                 android:textSize="24sp"/>
-        <io.github.karino2.kotlitex.MarkdownView
+        <io.github.karino2.kotlitex.view.MarkdownView
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:id="@+id/markdownView"
